chore: rework event bus class and interface
